Moved the business logic out of the AnnotationsContentViewer into its own class
