+++ Core/src/org/sleuthkit/autopsy/contentviewers/Bundle.properties-MERGED
@@ -15,22 +15,7 @@
 # governing permissions and limitations under the License.
 #
 
-AnnotationsContentViewer.centralRepositoryEntry.title=Central Repository Comments
-AnnotationsContentViewer.centralRepositoryEntryDataLabel.case=Case:
-AnnotationsContentViewer.centralRepositoryEntryDataLabel.comment=Comment:
-AnnotationsContentViewer.centralRepositoryEntryDataLabel.path=Path:
-AnnotationsContentViewer.centralRepositoryEntryDataLabel.type=Type:
-AnnotationsContentViewer.fileHitEntry.artifactCommentTitle=Artifact Comment
-AnnotationsContentViewer.fileHitEntry.comment=Comment:
-AnnotationsContentViewer.fileHitEntry.hashSetHitTitle=Hash Set Hit Comments
-AnnotationsContentViewer.fileHitEntry.interestingFileHitTitle=Interesting File Hit Comments
-AnnotationsContentViewer.fileHitEntry.setName=Set Name:
 AnnotationsContentViewer.onEmpty=No annotations were found for this particular item.
-AnnotationsContentViewer.sourceFile.title=Source File
-AnnotationsContentViewer.tagEntry.title=Tags
-AnnotationsContentViewer.tagEntryDataLabel.comment=Comment:
-AnnotationsContentViewer.tagEntryDataLabel.tag=Tag:
-AnnotationsContentViewer.tagEntryDataLabel.tagUser=Examiner:
 AnnotationsContentViewer.title=Annotations
 AnnotationsContentViewer.toolTip=Displays tags and comments associated with the selected content.
 ApplicationContentViewer.title=Application
