+++ Core/src/org/sleuthkit/autopsy/contentviewers/application/Bundle.properties-MERGED
@@ -1,0 +1,18 @@
+Annotations.centralRepositoryEntry.title=Central Repository Comments
+Annotations.centralRepositoryEntryDataLabel.case=Case:
+Annotations.centralRepositoryEntryDataLabel.comment=Comment:
+Annotations.centralRepositoryEntryDataLabel.path=Path:
+Annotations.centralRepositoryEntryDataLabel.type=Type:
+Annotations.fileHitEntry.artifactCommentTitle=Artifact Comment
+Annotations.fileHitEntry.comment=Comment:
+Annotations.fileHitEntry.hashSetHitTitle=Hash Set Hit Comments
+Annotations.fileHitEntry.interestingFileHitTitle=Interesting File Hit Comments
+Annotations.fileHitEntry.setName=Set Name:
+Annotations.onEmpty=No annotations were found for this particular item.
+Annotations.sourceFile.title=Source File
+Annotations.tagEntry.title=Tags
+Annotations.tagEntryDataLabel.comment=Comment:
+Annotations.tagEntryDataLabel.tag=Tag:
+Annotations.tagEntryDataLabel.tagUser=Examiner:
+Annotations.title=Annotations
+Annotations.toolTip=Displays tags and comments associated with the selected content.
